+++ src/compiler/config/test/validate-config.spec.ts
@@ -472,4 +472,23 @@ describe('validation', () => {
       expect(config.buildDist).toBe(config.buildEs5);
     });
   });
+
+  describe('validatePrimaryPackageOutputTarget', () => {
+    it('should default to false', () => {
+      const { config } = validateConfig(userConfig, bootstrapConfig);
+
+      expect(config.validatePrimaryPackageOutputTarget).toBe(false);
+    });
+
+    it.each([true, false])(
+      'should set validatePrimaryPackageOutputTarget to %p',
+      (validatePrimaryPackageOutputTarget) => {
+        userConfig.validatePrimaryPackageOutputTarget = validatePrimaryPackageOutputTarget;
+
+        const { config } = validateConfig(userConfig, bootstrapConfig);
+
+        expect(config.validatePrimaryPackageOutputTarget).toBe(validatePrimaryPackageOutputTarget);
+      }
+    );
+  });
 });